backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
openai>=1.52.0
pypdfium2>=5.0.0
Pillow>=10.0.0
httpx>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.9
beautifulsoup4>=4.12.0
requests>=2.32.0
tqdm>=4.66.0
python-dotenv>=1.0.0
